An R6 Class to represent a SBM

An R6 Class to represent a SBM

An R6 Class to represent a SBM

Active bindings

nNodes

The number of nodes

nBlocks

The number of blocks

nCovariates

The number of covariates

nDyads

The number of dyads

direction

character indicating if the network is directed or not

hasCovariates

boolean indicating if the SBM owns covariate(s)

mixtureParam

the vector of mixture parameters (block proportions)

connectParam

the matrix of connectivity: inter/intra probabilities of connection when the network does not have covariates, or a logit scaled version of it.

adjacencyMatrix

The adjacency matrix of the network

covarParam

the vector of parameters associated with the covariates

covarArray

the array of covariates

df_mixtureParams

degrees of freedoms for the mixture parameters

df_connectParams

degrees of freedoms for the connectivity parameters

df_covarParams

degrees of freedoms for the covariate parameters

Methods

Public methods

SBM$new()

SBM$show()

SBM$plot()

SBM$clone()

Method new()

Initialize an [`SBM`] object

Usage

SBM$new(
  directed = FALSE,
  nNodes = NA,
  mixtureParam = NA,
  connectParam = NA,
  covarParam = NULL,
  covarArray = NULL
)

Arguments

directed

logical indicating if the network is directed or not

nNodes

The number of nodes

mixtureParam

the vector of mixture parameters

connectParam

the matrix of connectivity: inter/intra probabilities of connection when the network does not have covariates, or a logit scaled version of it.

covarParam

the vector of parameters associated with the covariates

covarArray

the array of covariates

Method show()

show method

Usage

SBM$show(type = "Stochastic Block Model\n")

Arguments

type

character used to specify the type of SBM

Method plot()

basic matrix plot method for SBM object

Usage

SBM$plot(type = c("network", "connectivity"))

Arguments

type

character for the type of plot: either 'network' (true connection) or 'connectivity' (fitted connection). Default to 'network'.

Method clone()

The objects of this class are cloneable with this method.

Usage

SBM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.