An R6 Class to represent a sampler for a SBM

An R6 Class to represent a sampler for a SBM

An R6 Class to represent a sampler for a SBM

Details

The function [`simulate`] produces an instance of an object with class [`SBM_sampler`].

All fields of this class are only accessible for reading. This class comes with a set of methods, some of them being useful for the user (see examples)

R6 methods:$rBlocks(), $rAdjancencyMatrix()

S3 methodsprint(), plot()

Fields are accessed via active binding and cannot be changed by the user.

See also

Super class

missSBM::SBM -> SBM_sampler

Active bindings

blocks

matrix for clustering memberships

memberships

vector of clustering

connectProb

expected values of connection under the current model

Methods

Public methods

SBM_sampler$new()

SBM_sampler$rBlocks()

SBM_sampler$rAdjMatrix()

SBM_sampler$show()

SBM_sampler$clone()

Inherited methods

missSBM::SBM$plot()

Method new()

Initialize a [`SBM_sampler`] object

Usage

SBM_sampler$new(
  directed = FALSE,
  nNodes = NA,
  mixtureParam = NA,
  connectParam = NA,
  covarParam = NULL,
  covarArray = NULL
)

Arguments

directed

logical indicating if the network is directed or not

nNodes

The number of nodes

mixtureParam

the vector of mixture parameters

connectParam

the matrix of connectivity: inter/intra probabilities of connection when the network does not have covariates, or a logit scaled version of it.

covarParam

the vector of parameters associated with the covariates

covarArray

the array of covariates

Method rBlocks()

a method to generate a vector of clusters indicators

Usage

SBM_sampler$rBlocks()

Method rAdjMatrix()

a method to sample an adjacency matrix for the current SBM

Usage

SBM_sampler$rAdjMatrix()

Method show()

show method

Usage

SBM_sampler$show()

Method clone()

The objects of this class are cloneable with this method.

Usage

SBM_sampler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## SBM parameters
directed <- FALSE
N <- 300 # number of nodes
Q <- 3   # number of clusters
alpha <- rep(1,Q)/Q     # mixture parameter
pi <- diag(.45,Q) + .05 # connectivity matrix
gamma <- log(pi/(1-pi)) # logit transform fo the model with covariates
M <- 2 # two Gaussian covariates
covarMatrix <- matrix(rnorm(N*M,mean = 0, sd = 1), N, M)
covarParam  <- rnorm(M, -1, 1)

## draw a SBM without covariates through simulateSBM
sbm <- missSBM::simulate(N, alpha, pi, directed)

## equivalent construction from the SBM_sampler class itslef
sbm_s <- SBM_sampler$new(directed, N, alpha, pi)
sbm_s$rBlocks() # draw some blocks
sbm_s$rAdjMatrix() # draw some edges

coef(sbm_s, "mixture")

#> [1] 0.3333333 0.3333333 0.3333333

coef(sbm_s, "connectivity")

#>      [,1] [,2] [,3]
#> [1,] 0.50 0.05 0.05
#> [2,] 0.05 0.50 0.05
#> [3,] 0.05 0.05 0.50

summary(sbm_s)

#> Sampler for Stochastic Block Model
#> ==================================================================
#> Model undirected with 300 nodes, 3 blocks and no covariate(s).
#> ==================================================================
#> * Useful fields 
#>   $nNodes, $nBlocks, $nCovariates, $nDyads
#>   $mixtureParam, $connectParam
#>  $covarParam, $covarArray 
#> * Sampling methods 
#>   $rBlocks(), $rAdjmatrix()
#> * Additional fields 
#>   $blocks, $memberships, $adjMatrix, $connectProb
#> * S3 methods 
#>   plot, print, summary, coef