This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

Details

It is not designed not be call by the user

Super classes

sbm::SBM -> sbm::SBM_fit -> sbm::SimpleSBM_fit -> SimpleSBM_fit_missSBM

Active bindings

vExpec

double: variational approximation of the expectation complete log-likelihood

loglik

double: approximation of the log-likelihood (variational lower bound) reached

ICL

double: value of the integrated classification log-likelihood

Methods

Public methods

SimpleSBM_fit_missSBM$new()

SimpleSBM_fit_missSBM$doVEM()

SimpleSBM_fit_missSBM$update_parameters()

SimpleSBM_fit_missSBM$update_blocks()

SimpleSBM_fit_missSBM$clone()

Inherited methods

sbm::SBM$plot()

sbm::SBM$print()

sbm::SBM_fit$setModel()

sbm::SimpleSBM_fit$optimize()

sbm::SimpleSBM_fit$predict()

sbm::SimpleSBM_fit$reorder()

sbm::SimpleSBM_fit$show()

Method new()

constructor for simpleSBM_fit for missSBM purpose

Usage

SimpleSBM_fit_missSBM$new(adjacencyMatrix, clusterInit, covarList = list())

Arguments

adjacencyMatrix

a matrix encoding the graph

clusterInit

Initial clustering: either a character in "hierarchical", "spectral" or "kmeans", or a vector with size ncol(adjacencyMatrix), providing a user-defined clustering with nbBlocks levels. Default is "hierarchical".

covarList

An option list with M entries (the M covariates).

Method doVEM()

method to perform estimation via variational EM

Usage

SimpleSBM_fit_missSBM$doVEM(
  threshold = 1e-04,
  maxIter = 10,
  fixPointIter = 3,
  trace = FALSE
)

Arguments

threshold

stop when an optimization step changes the objective function by less than threshold. Default is 1e-4.

maxIter

V-EM algorithm stops when the number of iteration exceeds maxIter. Default is 10

fixPointIter

number of fix-point iterations in the Variational E step. Default is 3.

trace

logical for verbosity. Default is FALSE.

Method update_parameters()

update parameters estimation (M-step)

Usage

SimpleSBM_fit_missSBM$update_parameters()

Method update_blocks()

update variational estimation of blocks (VE-step)

Usage

SimpleSBM_fit_missSBM$update_blocks(log_lambda = 0)

Arguments

log_lambda

double use to adjust the parameter estimation according to the sampling design

Method clone()

The objects of this class are cloneable with this method.

Usage

SimpleSBM_fit_missSBM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.