Class for defining a double-standard sampler

Class for defining a double-standard sampler

Super classes

networkSampling -> networkSampler -> dyadSampler -> doubleStandardSampler

Methods

Public methods

doubleStandardSampler$new()

doubleStandardSampler$clone()

Inherited methods

networkSampling$print()

networkSampling$show()

networkSampler$rSamplingMatrix()

doubleStandardSampler$new()

constructor for networkSampling

Usage

doubleStandardSampler$new(parameters = NA, adjMatrix = NA, directed = FALSE)

Arguments

parameters

the vector of parameters associated to the sampling at play

adjMatrix

matrix of adjacency

directed

logical, directed network of not

doubleStandardSampler$clone()

The objects of this class are cloneable with this method.

Usage

doubleStandardSampler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.