Adjusting Stochastic Block Models under various missing data conditions

The missSBM package provides five functions:

simulatea function to define and draw network data according to a stochastic block model

samplea function to sample an existing network according to a variety of sampling designs

estimatea function to perform inference of SBM from network data with missing entries under various sampling designs.

smootha function to smooth an existing collection of SBM estimation, to avoid being trapped in local maxima.

Details

These function leads to the manipulation of a variety of R object, with their respective fields and methods. They are all automatically generated by the top-level functions itemized above, so that the user should generally to use their constructor or internal methods directly. The user should only have a basic understanding of the fields of each object to manipulate the output in R. The main objects are the following:

SBM_sampleran object to define a SBM to sample from

SBM_fitan object to define and store an SBM fit

networkSampleran object to define a network sampling to sample from

networkSamplingan object to define and store a network sampling fit

missSBM_fitan object that put together an SBM fit and and network sampling fit - the main point of the missSBM package !

missSBM_collectionan object to store a collection of missSBM_fit, ordered by number of block

References

Timothée Tabouy, Pierre Barbillon & Julien Chiquet (2019) “Variational Inference for Stochastic Block Models from Sampled Data”, Journal of the American Statistical Association, <doi:10.1080/01621459.2018.1562934>

Author

Timothée Tabouy, Pierre Barbillon, Julien Chiquet