missSBM: Handling Missing Data in Stochastic Block Models

When a network is partially observed (here, NAs in the adjacency matrix rather than 1 or 0 due to missing information between node pairs), it is possible to account for the underlying process that generates those NAs. 'missSBM', presented in 'Barbillon, Chiquet and Tabouy' (2022) doi:10.18637/jss.v101.i12 , adjusts the popular stochastic block model from network data sampled under various missing data conditions, as described in 'Tabouy, Barbillon and Chiquet' (2019) doi:10.1080/01621459.2018.1562934 .

The missSBM package provides the following top-level functions:

Details

observeNetwork a function to draw a partially observe network from an existing, fully observed network according to a variety of sampling designs

estimateMissSBM a function to perform inference of SBM from a partially observed under various sampling designs.

missSBM_param a function to define the list of parameters controlling estimateMissSBM

These function leads to the manipulation of a variety of R objects instantiated from some R6 classes, with their respective fields and methods. They are all generated by the top-level functions itemized above, so that the user should generally not use their constructor or internal methods directly. The user should only have a basic understanding of the fields of each object to manipulate the output in R. The main objects are the following:

missSBM_fit an object that put together an SBM fit and and network sampling fit - the main point of the missSBM package !

missSBM_collection an object to store a collection of missSBM_fit, ordered by number of block

SimpleSBM_fit_MNAR an object to define and store an SBM fit with MNAR values

SimpleSBM_fit_noCov an object to define and store an SBM fit without covariate, MAR values

SimpleSBM_fit_withCov an object to define and store an SBM fit with covariates, MAR values

networkSampling an object to define and store a network sampling fit

missSBM extends some functionality of the package sbm, by inheriting from classes and methods associated to simple stochastic block models.

References

Pierre Barbillon, Julien Chiquet & Timothée Tabouy (2022) "missSBM: An R Package for Handling Missing Values in the Stochastic Block Model", Journal of Statistical Software, doi:10.18637/jss.v101.i12

Timothée Tabouy, Pierre Barbillon & Julien Chiquet (2019) “Variational Inference for Stochastic Block Models from Sampled Data”, Journal of the American Statistical Association, doi:10.1080/01621459.2018.1562934

See also

Author

Maintainer: Julien Chiquet julien.chiquet@inrae.fr (ORCID)

Authors:

Julien Chiquet julien.chiquet@inrae.fr (ORCID)

Pierre Barbillon pierre.barbillon@agroparistech.fr (ORCID)

Timothée Tabouy timothee.tabouy@agroparistech.fr

Other contributors:

Jean-Benoist Léger jbleger@hds.utc.fr (provided C++ implementaion of K-means) [contributor]

François Gindraud francois.gindraud@gmail.com (provided C++ interface to NLopt) [contributor]

großBM team [contributor]

Pierre Barbillon pierre.barbillon@agroparistech.fr

Julien Chiquet julien.chiquet@inrae.fr

Timothée Tabouy timothee.tabouy@gmail.com