Prepare network data for estimation with missing data

This function puts together the adjacency matrix of a network and an optional list of covariates into a single sampledNetwork object, ready to use for inference with the estimate function of the missSBM package.

prepare_data(
  adjacencyMatrix,
  covariates = NULL,
  similarity = missSBM:::l1_similarity
)

Arguments

adjacencyMatrix
The adjacency matrix of the network (NAs allowed)

covariates
An optional list with M entries (the M covariates). If the covariates are node-centered, each entry of covariates must be a size-N vector; if the covariates are dyad-centered, each entry of covariates must be N x N matrix.

similarity
An optional R x R -> R function to compute similarities between node covariates. Default is l1_similarity, that is, -abs(x-y). Only relevant when the covariates is a list of size-N vectors.

Value

Returns an R6 object with class sampledNetwork.

See also

Examples

data(war)
adj_belligerent <- war$belligerent %>% igraph::as_adj(sparse = FALSE)
sampledNet_war_nocov <- prepare_data(adj_belligerent)
sampledNet_war_withcov <- prepare_data(adj_belligerent,
                             list(military_power = war$belligerent$power))