Simulatation of a Stochastic Block Model

Generates a realization (blocks and adjancency matrix) of a Stochastic Block model

simulateSBM(N, alpha, pi, directed = FALSE, covarMatrix = NULL,
  covarParam = NULL, covarSimilarity = l1_similarity)

Arguments

N
The number of nodes

alpha
The mixture parameters

pi
The connectivity matrix (either probabilities of inter and intra clusters )

directed
Boolean variable to indicate whether the network is directed or not. Default to FALSE.

covarMatrix
An optional matrix of covariates with dimension N x M (M covariates per node).

covarParam
An optional vector of parameters associated with the covariates, with size M

covarSimilarity
An optional R x R -> R function to compute similarity between node covariates. Default is l1_similarity, that is, -abs(x-y).

Value

an object with class SBM_sampler

See also

Examples

## SBM parameters
directed <- FALSE
N <- 300 # number of nodes
Q <- 3   # number of clusters
alpha <- rep(1,Q)/Q     # mixture parameter
pi <- diag(.45,Q) + .05 # connectivity matrix
gamma <- log(pi/(1-pi)) # logit transform fo the model with covariates
M <- 2 # two Gaussian covariates
covarMatrix <- matrix(rnorm(N*M,mean = 0, sd = 1), N, M)
covarParam  <- rnorm(M, -1, 1)

## draw a SBM without covariates
sbm <- simulateSBM(N, alpha, pi, directed)

## draw a SBM model with covariates
sbm_cov <- simulateSBM(N, alpha, gamma, directed, covarMatrix, covarParam)
# NOT RUN {
par(mfrow = c(1,2))
plot(sbm)
plot(sbm_cov)
# }