War_graphs data set

This dataset contains two networks where the nodes are countries and an edge in network "beligerent" means that the two countries have been at least once at war between years 1816 to 2007 while an edge in network "alliance" means that the two countries have had a formal alliance between years 1816 to 2012. The network `beligenrent` have less nodes since countries which have not been at war are not considered.

war_graphs

Format

A list with 2 two igraph objects, alliance and beligerent. Each grapha have two vertex attributes: 'name' (the country name) and 'military_power' (a score related to military power: the higer, the better)

Source

networs were extracted from <http://www.correlatesofwar.org/>

References

Sarkees, Meredith Reid and Frank Wayman (2010). Resort to War: 1816 - 2007. Washington DC: CQ Press.

Gibler, Douglas M. 2009. International military alliances, 1648-2008. CQ Press

Examples

data(war_graphs)
class(war_graphs$beligerent)

#> [1] "igraph"

igraph::gorder(war_graphs$alliance)

#> [1] 182

igraph::edges(war_graphs$alliance)

#> [[1]]
#> IGRAPH c364df0 UN-- 182 1694 -- 
#> + attr: name (v/c), military_power (v/n)
#> + edges from c364df0 (vertex names):
#>  [1] United States of America--Canada             
#>  [2] United States of America--Bahamas            
#>  [3] United States of America--Cuba               
#>  [4] United States of America--Haiti              
#>  [5] United States of America--Dominican Republic 
#>  [6] United States of America--Jamaica            
#>  [7] United States of America--Trinidad and Tobago
#>  [8] United States of America--Barbados           
#> + ... omitted several edges
#> 
#> attr(,"class")
#> [1] "igraph.edge"

igraph::get.vertex.attribute(war_graphs$alliance)

#> $name
#>   [1] "United States of America"         "Canada"                          
#>   [3] "Bahamas"                          "Cuba"                            
#>   [5] "Haiti"                            "Dominican Republic"              
#>   [7] "Jamaica"                          "Trinidad and Tobago"             
#>   [9] "Barbados"                         "Dominica"                        
#>  [11] "Grenada"                          "St. Lucia"                       
#>  [13] "St. Vincent and the Grenadines"   "Antigua & Barbuda"               
#>  [15] "St. Kitts and Nevis"              "Mexico"                          
#>  [17] "Belize"                           "Guatemala"                       
#>  [19] "Honduras"                         "El Salvador"                     
#>  [21] "Nicaragua"                        "Costa Rica"                      
#>  [23] "Panama"                           "Colombia"                        
#>  [25] "Venezuela"                        "Guyana"                          
#>  [27] "Suriname"                         "Ecuador"                         
#>  [29] "Peru"                             "Brazil"                          
#>  [31] "Bolivia"                          "Paraguay"                        
#>  [33] "Chile"                            "Argentina"                       
#>  [35] "Uruguay"                          "United Kingdom"                  
#>  [37] "Netherlands"                      "Belgium"                         
#>  [39] "Luxembourg"                       "France"                          
#>  [41] "Spain"                            "Portugal"                        
#>  [43] "Hanover"                          "Bavaria"                         
#>  [45] "Germany"                          "German Federal Republic"         
#>  [47] "German Democratic Republic"       "Baden"                           
#>  [49] "Saxony"                           "Wuerttemburg"                    
#>  [51] "Hesse Electoral"                  "Hesse Grand Ducal"               
#>  [53] "Mecklenburg Schwerin"             "Poland"                          
#>  [55] "Austria-Hungary"                  "Austria"                         
#>  [57] "Hungary"                          "Czechoslovakia"                  
#>  [59] "Czech Republic"                   "Slovakia"                        
#>  [61] "Italy"                            "Papal States"                    
#>  [63] "Two Sicilies"                     "Modena"                          
#>  [65] "Parma"                            "Tuscany"                         
#>  [67] "Malta"                            "Albania"                         
#>  [69] "Croatia"                          "Yugoslavia"                      
#>  [71] "Bosnia and Herzegovina"           "Greece"                          
#>  [73] "Cyprus"                           "Bulgaria"                        
#>  [75] "Moldova"                          "Romania"                         
#>  [77] "Russia"                           "Estonia"                         
#>  [79] "Latvia"                           "Lithuania"                       
#>  [81] "Ukraine"                          "Belarus"                         
#>  [83] "Armenia"                          "Georgia"                         
#>  [85] "Azerbaijan"                       "Finland"                         
#>  [87] "Sweden"                           "Norway"                          
#>  [89] "Denmark"                          "Iceland"                         
#>  [91] "Cape Verde"                       "Sao Tome and Principe"           
#>  [93] "Guinea-Bissau"                    "Equatorial Guinea"               
#>  [95] "Gambia"                           "Mali"                            
#>  [97] "Senegal"                          "Benin"                           
#>  [99] "Mauritania"                       "Niger"                           
#> [101] "Ivory Coast"                      "Guinea"                          
#> [103] "Burkina Faso"                     "Liberia"                         
#> [105] "Sierra Leone"                     "Ghana"                           
#> [107] "Togo"                             "Cameroon"                        
#> [109] "Nigeria"                          "Gabon"                           
#> [111] "Central African Republic"         "Chad"                            
#> [113] "Congo"                            "Democratic Republic of the Congo"
#> [115] "Uganda"                           "Kenya"                           
#> [117] "Tanzania"                         "Burundi"                         
#> [119] "Rwanda"                           "Somalia"                         
#> [121] "Djibouti"                         "Ethiopia"                        
#> [123] "Eritrea"                          "Angola"                          
#> [125] "Mozambique"                       "Zambia"                          
#> [127] "Zimbabwe"                         "South Africa"                    
#> [129] "Namibia"                          "Swaziland"                       
#> [131] "Madagascar"                       "Mauritius"                       
#> [133] "Morocco"                          "Algeria"                         
#> [135] "Tunisia"                          "Libya"                           
#> [137] "Sudan"                            "South Sudan"                     
#> [139] "Iran"                             "Turkey"                          
#> [141] "Iraq"                             "Egypt"                           
#> [143] "Syria"                            "Lebanon"                         
#> [145] "Jordan"                           "Israel"                          
#> [147] "Saudi Arabia"                     "Yemen Arab Republic"             
#> [149] "Yemen"                            "Yemen People's Republic"         
#> [151] "Kuwait"                           "Bahrain"                         
#> [153] "Qatar"                            "United Arab Emirates"            
#> [155] "Oman"                             "Afghanistan"                     
#> [157] "Turkmenistan"                     "Tajikistan"                      
#> [159] "Kyrgyzstan"                       "Uzbekistan"                      
#> [161] "Kazakhstan"                       "China"                           
#> [163] "Mongolia"                         "Taiwan"                          
#> [165] "Korea"                            "North Korea"                     
#> [167] "South Korea"                      "Japan"                           
#> [169] "India"                            "Pakistan"                        
#> [171] "Bangladesh"                       "Myanmar"                         
#> [173] "Thailand"                         "Cambodia"                        
#> [175] "Laos"                             "Vietnam"                         
#> [177] "Republic of Vietnam"              "Malaysia"                        
#> [179] "Philippines"                      "Indonesia"                       
#> [181] "Australia"                        "New Zealand"                     
#> 
#> $military_power
#>   [1] 1.580916e-01 1.154060e-02 1.005575e-04 1.952134e-03 4.768081e-04
#>   [6] 4.882688e-04 1.687000e-04 5.223961e-04 1.932553e-05 2.381714e-06
#>  [11] 4.306923e-06 4.686176e-06 3.533529e-06 3.127500e-06 1.788000e-06
#>  [16] 8.222487e-03 1.768438e-05 5.563076e-04 2.712289e-04 3.613949e-04
#>  [21] 2.744372e-04 1.602108e-04 1.987909e-04 2.426901e-03 2.329751e-03
#>  [26] 6.172979e-05 5.691053e-05 7.318107e-04 1.989205e-03 1.311592e-02
#>  [31] 6.895448e-04 3.972438e-04 2.332482e-03 5.258147e-03 8.469962e-04
#>  [36] 1.498935e-01 1.033064e-02 1.411523e-02 1.854537e-03 7.391118e-02
#>  [41] 1.790492e-02 4.619688e-03 3.211523e-03 7.480036e-03 8.148715e-02
#>  [46] 3.394109e-02 9.227503e-03 2.173479e-03 4.072688e-03 2.370157e-03
#>  [51] 1.495869e-03 1.717063e-03 6.665000e-04 1.478765e-02 6.143072e-02
#>  [56] 3.528368e-03 3.807748e-03 1.056389e-02 2.815900e-03 1.507535e-03
#>  [61] 2.330426e-02 7.303293e-03 2.429196e-02 6.910421e-04 6.826400e-04
#>  [66] 2.482033e-03 3.335918e-05 4.820674e-04 9.540524e-04 3.878271e-03
#>  [71] 5.593476e-04 2.600578e-03 1.763642e-04 2.710373e-03 4.497682e-04
#>  [76] 6.174967e-03 1.253533e-01 5.106711e-04 8.908422e-04 8.532756e-04
#>  [81] 1.395975e-02 2.244614e-03 6.538000e-04 5.287545e-04 1.343850e-03
#>  [86] 1.739672e-03 8.195443e-03 1.615947e-03 3.067616e-03 2.361884e-05
#>  [91] 2.765789e-05 9.853421e-06 1.211154e-04 4.587556e-05 4.641875e-05
#>  [96] 4.523792e-04 5.823642e-04 2.985000e-04 1.944321e-04 3.516887e-04
#> [101] 8.046736e-04 4.233927e-04 4.572547e-04 1.594688e-04 2.467577e-04
#> [106] 9.246411e-04 2.246057e-04 7.571906e-04 6.434683e-03 1.350491e-04
#> [111] 1.835868e-04 4.185849e-04 2.890264e-04 2.791419e-03 9.897471e-04
#> [116] 1.292794e-03 1.453488e-03 3.464020e-04 4.177667e-04 5.537226e-04
#> [121] 1.101639e-04 3.808574e-03 1.553195e-03 1.577871e-03 1.034192e-03
#> [126] 7.125959e-04 9.825375e-04 6.014670e-03 1.525957e-04 5.160444e-05
#> [131] 6.724811e-04 6.940222e-05 2.271024e-03 3.191206e-03 1.726150e-03
#> [136] 1.188921e-03 1.805128e-03 1.124100e-03 6.503381e-03 2.248162e-02
#> [141] 4.199323e-03 7.152559e-03 2.876909e-03 5.561104e-04 9.179149e-04
#> [146] 2.755817e-03 4.895434e-03 5.867400e-04 1.510165e-03 3.827167e-04
#> [151] 1.580642e-03 4.568810e-04 6.802357e-04 1.742474e-03 7.699905e-04
#> [156] 1.858001e-03 7.121409e-04 3.686818e-04 3.687273e-04 2.671255e-03
#> [161] 3.442377e-03 1.331898e-01 3.190707e-04 7.702361e-03 2.891295e-03
#> [166] 8.269331e-03 1.494973e-02 3.854824e-02 5.832334e-02 1.137814e-02
#> [171] 6.200181e-03 3.840391e-03 4.142872e-03 1.141755e-03 5.585200e-04
#> [176] 7.547656e-03 5.667073e-03 2.679616e-03 4.410028e-03 1.335477e-02
#> [181] 7.507588e-03 8.032839e-04
#>