An R6 Class to represent a sampler for a SBM

An R6 Class to represent a sampler for a SBM

An R6 Class to represent a sampler for a SBM

Details

The function [`simulate`] produces an instance of an object with class [`SBM_sampler`].

All fields of this class are only accessible for reading. This class comes with a set of methods, some of them being useful for the user (see examples)

R6 methods:$rMemberships(), $rAdjacency()

S3 methodsprint(), plot()

Fields are accessed via active binding and cannot be changed by the user.

See also

Super class

missSBM::SBM -> SBM_sampler

Active bindings

indMemberships

matrix for clustering memberships

memberships

vector of clustering

expectation

expected values of connection under the current model

Methods

Public methods

SBM_sampler$new()

SBM_sampler$rMemberships()

SBM_sampler$rAdjacency()

SBM_sampler$show()

SBM_sampler$print()

SBM_sampler$clone()

Inherited methods

missSBM::SBM$plot()

Method new()

Initialize a [`SBM_sampler`] object

Usage

SBM_sampler$new(
  directed = FALSE,
  nbNodes = NA,
  blockProp = NA,
  connectParam = NA,
  covarParam = numeric(0),
  covarList = list()
)

Arguments

directed

logical indicating if the network is directed or not

nbNodes

The number of nodes

blockProp

the vector of mixture parameters

connectParam

the matrix of connectivity: inter/intra probabilities of connection when the network does not have covariates, or a logit scaled version of it.

covarParam

the vector of parameters associated with the covariates

covarList

A list with M entries (the M covariates). Each entry of the list must be an N x N matrix

Method rMemberships()

a method to generate a vector of clusters indicators

Usage

SBM_sampler$rMemberships()

Method rAdjacency()

a method to sample an adjacency matrix for the current SBM

Usage

SBM_sampler$rAdjacency()

Method show()

show method

Usage

SBM_sampler$show()

Method print()

User friendly print method

Usage

SBM_sampler$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

SBM_sampler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## SBM parameters
directed <- FALSE
N <- 300 # number of nodes
Q <- 3   # number of clusters
pi <- rep(1,Q)/Q     # mixture parameter
theta <- diag(.45,Q) + .05 # connectivity matrix

## draw a SBM without covariates through simulateSBM
sbm <- missSBM::simulate(N, pi, theta, directed)

## equivalent construction from the SBM_sampler class itslef
sbm_s <- SBM_sampler$new(directed, N, pi, theta)
sbm_s$rMemberships() # draw some blocks
sbm_s$rAdjacency() # draw some edges

coef(sbm_s, "block")

#> [1] 0.3333333 0.3333333 0.3333333

coef(sbm_s, "connectivity")

#>      [,1] [,2] [,3]
#> [1,] 0.50 0.05 0.05
#> [2,] 0.05 0.50 0.05
#> [3,] 0.05 0.05 0.50

summary(sbm_s)

#> Sampler for Stochastic Block Model
#> ==================================================================
#> Model undirected with 300 nodes, 3 blocks and no covariate(s).
#> ==================================================================
#> * Useful fields 
#>   $nbNodes, $nbBlocks, $nbCovariates, $nbDyads
#>   $blockProp, $connectParam
#>  $covarParam, $covarArray 
#> * Sampling methods 
#>   $rMemberships(), $rAdjacency()
#> * Additional fields 
#>   $indMemberships, $memberships, $adjMatrix, $expectation
#> * S3 methods 
#>   plot, print, summary, coef