This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

Details

It is not designed not be call by the user

Super classes

sbm::SBM -> sbm::SimpleSBM -> SimpleSBM_fit

Active bindings

type

the type of SBM (distribution of edges values, network type, presence of covariates)

penalty

double, value of the penalty term in ICL

entropy

double, value of the entropy due to the clustering distribution

loglik

double: approximation of the log-likelihood (variational lower bound) reached

ICL

double: value of the integrated classification log-likelihood

Methods

Public methods

SimpleSBM_fit$new()

SimpleSBM_fit$doVEM()

SimpleSBM_fit$reorder()

SimpleSBM_fit$clone()

Inherited methods

sbm::SBM$print()

sbm::SBM$rNetwork()

sbm::SimpleSBM$plot()

sbm::SimpleSBM$predict()

sbm::SimpleSBM$rEdges()

sbm::SimpleSBM$rMemberships()

sbm::SimpleSBM$show()

Method new()

constructor for simpleSBM_fit for missSBM purpose

Usage

SimpleSBM_fit$new(networkData, clusterInit, covarList = list())

Arguments

networkData

a structure to store network under missing data condition: either a matrix possibly with NA, or a missSBM:::partlyObservedNetwork

clusterInit

Initial clustering: a vector with size ncol(adjacencyMatrix), providing a user-defined clustering with nbBlocks levels.

covarList

An optional list with M entries (the M covariates).

Method doVEM()

method to perform estimation via variational EM

Usage

SimpleSBM_fit$doVEM(
  threshold = 0.01,
  maxIter = 100,
  fixPointIter = 3,
  trace = FALSE
)

Arguments

threshold

stop when an optimization step changes the objective function by less than threshold. Default is 1e-4.

maxIter

V-EM algorithm stops when the number of iteration exceeds maxIter. Default is 10

fixPointIter

number of fix-point iterations in the Variational E step. Default is 5.

trace

logical for verbosity. Default is FALSE.

Method reorder()

permute group labels by order of decreasing probability

Usage

SimpleSBM_fit$reorder()

Method clone()

The objects of this class are cloneable with this method.

Usage

SimpleSBM_fit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.