Internal class for a binary SBM fit under MNAR sampling designs (double-standard, block-node, block-dyad).

It is not designed to be called directly by the user.

Super classes

sbm::SBM -> sbm::SimpleSBM -> SimpleSBM_fit -> SimpleSBM_fit_noCov -> SimpleSBM_MNAR_noCov

Active bindings

vExpec

double: variational approximation of the expectation complete log-likelihood

Methods

Public methods

SimpleSBM_MNAR_noCov$new()

SimpleSBM_MNAR_noCov$update_parameters()

SimpleSBM_MNAR_noCov$update_blocks()

SimpleSBM_MNAR_noCov$polish_log_tau()

SimpleSBM_MNAR_noCov$clone()

Inherited methods

sbm::SBM$print()

sbm::SBM$rNetwork()

sbm::SimpleSBM$plot()

sbm::SimpleSBM$predict()

sbm::SimpleSBM$rEdges()

sbm::SimpleSBM$rMemberships()

sbm::SimpleSBM$show()

SimpleSBM_fit$doVEM()

SimpleSBM_fit$get_state()

SimpleSBM_fit$reorder()

SimpleSBM_fit$set_state()

SimpleSBM_MNAR_noCov$new()

constructor for simpleSBM_fit for missSBM purpose

Usage

SimpleSBM_MNAR_noCov$new(networkData, clusterInit)

Arguments

networkData

a structure to store network under missing data condition: either a matrix possibly with NA, or a missSBM:::partlyObservedNetwork

clusterInit

Initial clustering: a vector with size ncol(adjacencyMatrix), providing a user-defined clustering with nbBlocks levels.

SimpleSBM_MNAR_noCov$update_parameters()

update parameters estimation (M-step)

Usage

SimpleSBM_MNAR_noCov$update_parameters(nu = NULL)

Arguments

nu

currently imputed values

SimpleSBM_MNAR_noCov$update_blocks()

update variational estimation of blocks (VE-step)

Usage

SimpleSBM_MNAR_noCov$update_blocks(log_lambda = 0)

Arguments

log_lambda

additional term sampling dependent used to de-bias estimation of tau

SimpleSBM_MNAR_noCov$polish_log_tau()

for each node, the complete-data log-likelihood it would contribute to each class if hard-assigned there (theta/pi held fixed), used to decide node-swap moves in missSBM_fit$polish(). Unlike update_blocks()'s tau update (which combines the observed- and imputed-part E-steps as-is), this subtracts one copy of log(pi): both E-step calls add it, so summing them double-counts it relative to vExpec's convention – verified numerically to matter (see git history).

Usage

SimpleSBM_MNAR_noCov$polish_log_tau(log_lambda = 0)

Arguments

log_lambda

additional sampling-design-dependent term, added as-is

Returns

an N x Q matrix

SimpleSBM_MNAR_noCov$clone()

The objects of this class are cloneable with this method.

Usage

SimpleSBM_MNAR_noCov$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.