This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

Details

It is not designed not be call by the user

Super classes

sbm::SBM -> sbm::SimpleSBM -> missSBM::SimpleSBM_fit -> missSBM::SimpleSBM_fit_noCov -> SimpleSBM_MNAR_noCov

Active bindings

imputation

the matrix of imputed values

vExpec

double: variational approximation of the expectation complete log-likelihood

Methods

Public methods

SimpleSBM_fit_MNAR$new()

SimpleSBM_fit_MNAR$update_parameters()

SimpleSBM_fit_MNAR$update_blocks()

SimpleSBM_fit_MNAR$clone()

Inherited methods

sbm::SBM$print()

sbm::SBM$rNetwork()

sbm::SimpleSBM$plot()

sbm::SimpleSBM$predict()

sbm::SimpleSBM$rEdges()

sbm::SimpleSBM$rMemberships()

sbm::SimpleSBM$show()

missSBM::SimpleSBM_fit$doVEM()

missSBM::SimpleSBM_fit$reorder()

Method new()

constructor for simpleSBM_fit for missSBM purpose

Usage

SimpleSBM_fit_MNAR$new(networkData, clusterInit)

Arguments

networkData

a structure to store network under missing data condition: either a matrix possibly with NA, or a missSBM:::partlyObservedNetwork

clusterInit

Initial clustering: a vector with size ncol(adjacencyMatrix), providing a user-defined clustering with nbBlocks levels.

Method update_parameters()

update parameters estimation (M-step)

Usage

SimpleSBM_fit_MNAR$update_parameters(nu = NULL)

Arguments

nu

currently imputed values

Method update_blocks()

update variational estimation of blocks (VE-step)

Usage

SimpleSBM_fit_MNAR$update_blocks(log_lambda = 0)

Arguments

log_lambda

additional term sampling dependent used to de-bias estimation of tau

Method clone()

The objects of this class are cloneable with this method.

Usage

SimpleSBM_fit_MNAR$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.