This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

Details

It is not designed not be call by the user

Super classes

sbm::SBM -> sbm::SimpleSBM -> missSBM::SimpleSBM_fit -> SimpleSBM_fit_withCov

Active bindings

imputation

the matrix of imputed values

vExpec

double: variational approximation of the expectation complete log-likelihood

vExpec_corrected

double: variational approximation of the expectation complete log-likelihood with correction to be comparable with NMAR criteria

Methods

Public methods

SimpleSBM_fit_withCov$update_parameters()

SimpleSBM_fit_withCov$update_blocks()

SimpleSBM_fit_withCov$clone()

Inherited methods

sbm::SBM$print()

sbm::SBM$rNetwork()

sbm::SimpleSBM$plot()

sbm::SimpleSBM$predict()

sbm::SimpleSBM$rEdges()

sbm::SimpleSBM$rMemberships()

sbm::SimpleSBM$show()

missSBM::SimpleSBM_fit$doVEM()

missSBM::SimpleSBM_fit$initialize()

missSBM::SimpleSBM_fit$reorder()

Method update_parameters()

update parameters estimation (M-step)

Usage

SimpleSBM_fit_withCov$update_parameters(...)

Arguments

...

use for compatibility

control

a list to tune nlopt for optimization, see documentation of nloptr

Method update_blocks()

update variational estimation of blocks (VE-step)

Usage

SimpleSBM_fit_withCov$update_blocks(...)

Arguments

...

use for compatibility

Method clone()

The objects of this class are cloneable with this method.

Usage

SimpleSBM_fit_withCov$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.