Inference of an SBM with missing data

Perform variational inference of a Stochastic Block Model from a sampled adjacency matrix

estimate(adjacencyMatrix, vBlocks, sampling,
  clusterInit = ifelse(is.null(covarMatrix), "hierarchical", "spectral"),
  covarMatrix = NULL, covarSimilarity = l1_similarity, trace = TRUE,
  cores = 1, control_VEM = list())

Arguments

adjacencyMatrix
The adjacency matrix of the network

vBlocks
The vector of number of blocks considered in the collection

sampling
The sampling design for missing data modeling : "dyad", "node", "double-standard", "block-dyad", "block-node" ,"degree

clusterInit
Initial method for clustering: either a character in "hierarchical", "spectral" or "kmeans", or a list with length(vBlocks) vectors, each with size ncol(adjacencyMatrix) providing a user-defined clustering

covarMatrix
An optional matrix of covariates with dimension N x M (M covariates per node).

covarSimilarity
An optional R x R -> R function to compute similarity between node covariates. Default is #'

trace
logical, control the verbosity. Default to TRUE.

cores
integer, the number of cores to use when multiply model are fitted

control_VEM
a list controlling the variational EM algorithm. See details.

Value

Returns an R6 object with class missSBM_collection.

See also

Examples

## SBM parameters
directed <- FALSE
N <- 300 # number of nodes
Q <- 3   # number of clusters
alpha <- rep(1,Q)/Q     # mixture parameter
pi <- diag(.45,Q) + .05 # connectivity matrix

## simulate a SBM without covariates
sbm <- missSBM::simulate(N, alpha, pi, directed)

## Sample network data
samplingParameters <- .5 # the sampling rate
sampling <- "dyad"       # the sampling design
sampledNet <- missSBM::sample(sbm$adjMatrix, sampling, samplingParameters)

## Inference :
vBlocks <- 1:5 # number of classes
collection <- missSBM::estimate(sampledNet$adjMatrix, vBlocks, sampling, trace = FALSE)
collection$ICL

#> [1] 107404.94 101010.79  95767.52  95934.17  96122.00