Adjusting Stochastic Block Models under various missing data conditions

The missSBM package provides the following top-level functions functions:

observeNetworka function to draw a partially observe network from an existing, fully observed network according to a variety of sampling designs

estimateMissSBMa function to perform inference of SBM from a partially observed under various sampling designs.

smootha function to smooth an existing collection of missSBM estimation, to avoid being trapped in local maxima.

Details

These function leads to the manipulation of a variety of R objects instantiated from some R6 classes, with their respective fields and methods. They are all generated by the top-level functions itemized above, so that the user should generally not use their constructor or internal methods directly. The user should only have a basic understanding of the fields of each object to manipulate the output in R. The main objects are the following:

missSBM_fitan object that put together an SBM fit and and network sampling fit - the main point of the missSBM package !

missSBM_collectionan object to store a collection of missSBM_fit, ordered by number of block

SimpleSBM_fit_missSBMan object to define and store an SBM fit (with no missing values)

networkSamplingan object to define and store a network sampling fit

missSBM extends some functionality of the package sbm, by inheriting from classes and methods associated to simple stochastic block models.

References

Timothée Tabouy, Pierre Barbillon & Julien Chiquet (2019) “Variational Inference for Stochastic Block Models from Sampled Data”, Journal of the American Statistical Association, doi: 10.1080/01621459.2018.1562934

Author

Timothée Tabouy, Pierre Barbillon, Julien Chiquet