An object to represent a collection of missSBM_fit

An object to represent a collection of missSBM_fit

An object to represent a collection of missSBM_fit

Details

This R6 class stores a collection of missSBM_fit. Comes with basic printing methods and field access.

See also

Active bindings

models

a list of models

ICL

the vector of Integrated Classification Criterion (ICL) associated to the models in the collection (the smaller, the better)

bestModel

the best model according to the ICL

vBlocks

a vector with the number of blocks

optimizationStatus

a data.frame summarizing the optimization process for all models

Methods

Public methods

missSBM_collection$new()

missSBM_collection$estimate()

missSBM_collection$smooth()

missSBM_collection$show()

missSBM_collection$clone()

Method new()

constructor for networkSampling

Usage

missSBM_collection$new(
  sampledNet,
  vBlocks,
  sampling,
  clusterInit,
  cores,
  trace,
  useCov
)

Arguments

sampledNet

An object with class [`sampledNetwork`], typically obtained with the function [`prepare_data`] (real-word data) or [`sample`] (simulation).

vBlocks

vector of integer with the number of blocks in the successively fitted models

sampling

The sampling design for the modelling of missing data: MAR designs ("dyad", "node") and NMAR designs ("double-standard", "block-dyad", "block-node" ,"degree")

clusterInit

Initial method for clustering: either a character in "hierarchical", "spectral" or "kmeans", or a list with length(vBlocks) vectors, each with size ncol(adjacencyMatrix), providing a user-defined clustering. Default is "hierarchical".

cores

integer for number of cores used. Default is 1.

trace

integer for verbosity (0, 1, 2). Default is 1. Useless when cores > 1

useCov

logical. If covariates are present in sampledNet, should they be used for the inference or of the network sampling design, or just for the SBM inference? default is TRUE.

Method estimate()

method to launch the estimation of the collection of models

Usage

missSBM_collection$estimate(control)

Arguments

control

a list of parameters controlling the variational EM algorithm. See details of function [`estimate`]

Method smooth()

method for performing smoothing of the ICL

Usage

missSBM_collection$smooth(type, control)

Arguments

type

character, the type of smoothing: forward, backward, both

control

a list of parameters controlling the smoothing. See details of regular function [`smooth`]

Method show()

show method for missSBM_collection

Usage

missSBM_collection$show()

Method clone()

The objects of this class are cloneable with this method.

Usage

missSBM_collection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.