Class for defining a simple dyad sampler

Class for defining a simple dyad sampler

Class for defining a simple dyad sampler

Super classes

missSBM::networkSampling -> missSBM::networkSampler -> missSBM::dyadSampler -> simpleDyadSampler

Methods

Public methods

simpleDyadSampler$new()

simpleDyadSampler$clone()

Inherited methods

missSBM::networkSampling$print()

missSBM::networkSampling$show()

missSBM::networkSampler$rSamplingMatrix()

Method new()

constructor for networkSampling

Usage

simpleDyadSampler$new(
  parameters = NA,
  nbNodes = NA,
  directed = FALSE,
  covarArray = NULL,
  intercept = 0
)

Arguments

parameters

the vector of parameters associated to the sampling at play

nbNodes

number of nodes in the network

directed

logical, directed network of not

covarArray

an array of covariates used

intercept

double, intercept term used to compute the probability of sampling in the presence of covariates. Default 0.

Method clone()

The objects of this class are cloneable with this method.

Usage

simpleDyadSampler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.