Class for defining a simple node sampler

Class for defining a simple node sampler

Class for defining a simple node sampler

Super classes

missSBM::networkSampling -> missSBM::networkSampler -> missSBM::nodeSampler -> simpleNodeSampler

Methods

Public methods

simpleNodeSampler$new()

simpleNodeSampler$clone()

Inherited methods

missSBM::networkSampling$show()

missSBM::networkSampler$rSamplingMatrix()

Method new()

constructor for networkSampling

Usage

simpleNodeSampler$new(
  parameters = NA,
  nNodes = NA,
  directed = FALSE,
  covarMatrix = NULL,
  intercept = 0
)

Arguments

parameters

the vector of parameters associated to the sampling at play

nNodes

number of nodes in the network

directed

logical, directed network of not

covarMatrix

a matrix of covariates used

intercept

double, intercept term used to compute the probability of sampling in the presence of covariates. Default 0.

Method clone()

The objects of this class are cloneable with this method.

Usage

simpleNodeSampler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.